requirements.txt
aioredis==2.0.1
amqp==5.1.1
arq==0.25.0
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.1
Django==4.1.7
django-cors-headers==4.0.0
django-filter==23.2
django-rest-knox==4.2.0
django-slugify-processor==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
hiredis==2.2.2
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
schedule==1.2.0
dadata==21.10.1
elasticsearch>=7,<8
elasticsearch-dsl>=7,<8
google-auth==2.22.0
python-memcached==1.59
pymemcache==4.0.0
httpx==0.25.0
boto3
json2xml
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
